requirements.txt
Pillow
einops
safetensors
timm
tomesd
torch>=2.1.2
torchdiffeq
torchsde
decord
datasets
numpy
scikit-image
opencv-python
omegaconf
SentencePiece
albumentations
imageio[ffmpeg]
imageio[pyav]
tensorboard
beautifulsoup4
ftfy
func_timeout
onnxruntime
accelerate>=0.25.0
gradio>=3.41.2
diffusers>=0.30.1
transformers>=4.46.2
huggingface-hub==0.31.2
tokenizers==0.20.3

# qwen3-vl
pip install git+https://github.com/huggingface/transformers
# qwen-image  
pip install git+https://github.com/huggingface/diffusers

# 恢复
pip install  transformers==4.46.2 huggingface-hub==0.31.2 tokenizers==0.20.3